Revert "fix(dropping): fixes various itemdrop-related issues"
This reverts commit fe0683a1912e3268ddf2ee6ed16cc03a359f19ec.

+++ code/game/gamemodes/changeling/helpcode/items.dm
@@ -14,8 +14,6 @@
 	throwforce = 0 //Just to be on the safe side
 	throw_range = 0
 	throw_speed = 0
-	canremove = FALSE
-	force_drop = TRUE
 	var/mob/living/creator //This is just like ninja swords, needed to make sure dumb shit that removes the sword doesn't make it stay around.
 	var/weapType = "weapon"
 	var/weapLocation = "arm"
@@ -36,9 +34,10 @@
 	user.visible_message(SPAN("danger", "With a sickening crunch, [creator] reforms their arm!"), \
 						 SPAN("changeling", "We assimilate the weapon back into our body."), \
 						 SPAN("italics", "You hear organic matter ripping and tearing!"))
-	playsound(user, 'sound/effects/blob/blobattack.ogg', 30, 1)
+	playsound(src, 'sound/effects/blob/blobattack.ogg', 30, 1)
 	spawn(1)
-		qdel(src)
+		if(src)
+			qdel(src)
 
 /obj/item/weapon/melee/changeling/Destroy()
 	STOP_PROCESSING(SSobj, src)
@@ -58,7 +57,8 @@
 			host.embedded -= src
 			host.drop_from_inventory(src)
 		spawn(1)
-			qdel(src)
+			if(src)
+				qdel(src)
 
 /obj/item/weapon/melee/changeling/arm_blade
 	name = "arm blade"
@@ -68,6 +68,8 @@
 	armor_penetration = 15
 	sharp = 1
 	edge = 1
+	anchored = 1
+	canremove = 0
 	attack_verb = list("attacked", "slashed", "stabbed", "sliced", "torn", "ripped", "diced", "cut")
 
 /obj/item/weapon/melee/changeling/arm_blade/greater
@@ -84,6 +86,7 @@
 	force = 15
 	sharp = 1
 	edge = 1
+	canremove = 0
 	attack_verb = list("attacked", "slashed", "stabbed", "sliced", "torn", "ripped", "diced", "cut")
 
 /obj/item/weapon/melee/changeling/claw/greater
@@ -91,6 +94,7 @@
 	desc = "A grotesque blade made out of bone and flesh that cleaves through people and armor as a hot knife through butter."
 	force = 20
 	armor_penetration = 50
+	anchored = 1
 
 ///// Changeling emag /////
 /obj/item/weapon/finger_lockpick
@@ -98,8 +102,6 @@
 	desc = "This finger appears to be an organic datajack."
 	icon = 'icons/obj/weapons.dmi'
 	icon_state = "electric_hand"
-	canremove = FALSE
-	force_drop = TRUE
 
 /obj/item/weapon/finger_lockpick/New()
 	if(ismob(loc))
@@ -108,7 +110,8 @@
 /obj/item/weapon/finger_lockpick/dropped(mob/user)
 	to_chat(user, SPAN("changeling", "We discreetly shape our finger back to a less suspicious form."))
 	spawn(1)
-		qdel(src)
+		if(src)
+			qdel(src)
 
 /obj/item/weapon/finger_lockpick/afterattack(atom/target, mob/living/user, proximity)
 	if(!target)
